requirements.txt
torch
torchvision
numpy
scikit-learn
pillow
matplotlib
seaborn
